Streamline "fields updated" events
`FieldsUpdatedEvent` (renamed to `FieldsUpdatedPersistentEvent`) now extends
`EntityUpdatedPersistentEvent` so event classes don't have to implement both.

There are new convenience methods for constructing `UpdatedField` instances
to make `listUpdatedFields` implementations more concise.

`FieldsUpdatedPersistentEvent` includes additional documentation to help
clarify how to implement it.
