+++ src/main/kotlin/com/terraformation/backend/eventlog/EntityEvents.kt
@@ -12,5 +12,9 @@ interface EntityCreatedPersistentEvent : PersistentEvent
  */
 interface EntityDeletedPersistentEvent : PersistentEvent
 
-/** Marker interface for entity update events. */
+/**
+ * Marker interface for entity update events. If possible, events should implement
+ * [FieldsUpdatedPersistentEvent] which extends this interface. Classes that implement this
+ * interface but not that one must be explicitly handled in [EventLogPayloadTransformer.getActions].
+ */
 interface EntityUpdatedPersistentEvent : PersistentEvent
