+++ src/main/kotlin/com/terraformation/backend/tracking/event/Events.kt
@@ -13,8 +13,7 @@ import com.terraformation.backend.db.tracking.PlantingSiteId
 import com.terraformation.backend.db.tracking.PlantingZoneId
 import com.terraformation.backend.eventlog.EntityCreatedPersistentEvent
 import com.terraformation.backend.eventlog.EntityDeletedPersistentEvent
-import com.terraformation.backend.eventlog.EntityUpdatedPersistentEvent
-import com.terraformation.backend.eventlog.FieldsUpdatedEvent
+import com.terraformation.backend.eventlog.FieldsUpdatedPersistentEvent
 import com.terraformation.backend.eventlog.PersistentEvent
 import com.terraformation.backend.ratelimit.RateLimitedEvent
 import com.terraformation.backend.tracking.edit.PlantingSiteEdit
@@ -267,13 +266,13 @@ data class ObservationMediaFileEditedEventV1(
     override val observationId: ObservationId,
     override val organizationId: OrganizationId,
     override val plantingSiteId: PlantingSiteId,
-) : EntityUpdatedPersistentEvent, FieldsUpdatedEvent, ObservationMediaFilePersistentEvent {
+) : FieldsUpdatedPersistentEvent, ObservationMediaFilePersistentEvent {
   data class Values(
       val caption: String?,
   )
 
   override fun listUpdatedFields() =
-      listOf(FieldsUpdatedEvent.UpdatedField("caption", changedFrom.caption, changedTo.caption))
+      listOfNotNull(createUpdatedField("caption", changedFrom.caption, changedTo.caption))
 }
 
 typealias ObservationMediaFileEditedEvent = ObservationMediaFileEditedEventV1
